Trainer File added (Easiest train method and fastest :) )

+++ utils/utils.py
@@ -92,7 +92,7 @@ def detokenize_words(word: list, first_word_token: int = 0, last_word_token: int
     return w
 
 
-def count_model_parameters(model, div: int = 1e6):
+def count_model_parameters(model, div: float = 1e6):
     return sum(m.numel() for m in model.parameters()) / div
